+++ backend/core/src/listings/listings.service.spec.ts
@@ -5,6 +5,7 @@ import { HttpException, HttpStatus } from "@nestjs/common"
 import { Listing } from "./entities/listing.entity"
 import { ListingsQueryParams, ListingFilterParams } from "./dto/listing.dto"
 import { Compare } from "../shared/dto/filter.dto"
+import { TranslationsService } from "../translations/translations.service"
 
 // Cypress brings in Chai types for the global expect, but we want to use jest
 // expect here so we need to re-declare it.
@@ -100,6 +101,10 @@ describe("ListingsService", () => {
         {
           provide: getRepositoryToken(Listing),
           useValue: mockListingsRepo,
+        },
+        {
+          provide: TranslationsService,
+          useValue: { translateListing: jest.fn() },
         },
       ],
     }).compile()