+++ sites/public/pages/listing/[id]/[slug].tsx
@@ -64,8 +64,10 @@ export async function getStaticPaths(context: { locales: Array<string> }) {
   }
 }
 
-export async function getStaticProps(context: { params: Record<string, string> }) {
-  const response = await axios.get(`${process.env.backendApiBase}/listings/${context.params.id}`)
+export async function getStaticProps(context: { params: Record<string, string>; locale: string }) {
+  const response = await axios.get(`${process.env.backendApiBase}/listings/${context.params.id}`, {
+    headers: { language: context.locale },
+  })
 
   return {
     props: {